+++ Campaign/resources/assets/js/components/PageviewRules.vue
@@ -1,44 +1,89 @@
 <template>
-    <div class="input-group m-t-20">
-        <span class="input-group-addon pageview-rules-addon"><i class="zmdi zmdi-eye"></i></span>
-        <div>
-            <div class="row">
-                <div class="col-md-12 col-sm-12 col-xs-12">
-                    <label class="fg-label">Pageview rules</label>
-                </div>
+    <div>
+        <div class="input-group m-t-20">
+            <span class="input-group-addon pageview-rules-addon"><i class="zmdi zmdi-eye"></i></span>
+            <div>
+                <div class="row">
+                    <div class="col-md-12 col-sm-12 col-xs-12">
+                        <label class="fg-label">Display banner</label>
+                    </div>
 
-                <div class="col-md-8 col-sm-8 col-xs-10 pageview-rules-wrapper">
-                    <div class="row" v-for="(pageviewRule, i) in pageviewRulesData">
-                        <div class="col-md-4 col-sm-4 col-xs-4">
-                            <v-select v-model="pageviewRule.rule"
-                                    id="rule"
-                                    :name="'pageview_rules[' + i + '][rule]'"
-                                    :value="pageviewRule.rule"
-                                    :options.sync="pageviewRulesOptions"
-                                    placeholder="Rule"
-                                    :title="'Rule'"
-                                    ref="pageviewRulesDataRefs"
-                            ></v-select>
-                        </div>
-                        <div class="col-md-3 col-sm-3 col-xs-3">
-                            <input v-model="pageviewRule.num" placeholder="n-th" class="form-control fg-input" :name="'pageview_rules[' + i + '][num]'" id="num" type="text">
-                        </div>
-                        <div class="col-md-3 col-sm-3 col-xs-3">
-                            <span style="line-height: 31px;">request</span>
-                        </div>
+                    <div class="col-md-8 col-sm-8 col-xs-10 pageview-rules-wrapper">
+                        <div class="row">
 
-                        <div class="col-md-2 col-sm-2 col-xs-2">
-                            <span class="btn btn-sm palette-Grey-400 bg waves-effect" v-on:click="removeRule(i)"><i class="zmdi zmdi-minus-square"></i></span>
-                        </div>
+                            <div class="input-group fg-float m-t-10 radio">
+                                <label class="m-l-15">
+                                    Always
+                                    <input v-model="displayBanner"
+                                           type="radio"
+                                           name="pageview_rules[display_banner]"
+                                           value="always">
+                                    <i class="input-helper"></i>
+                                </label>
+                            </div>
 
+                            <div class="col-md-4 col-sm-4" style="margin-left: -15px;">
+                                <div class="input-group fg-float m-t-10 radio">
+                                    <label class="m-l-15">
+                                        Every
+                                        <input v-model="displayBanner"
+                                               type="radio"
+                                               name="pageview_rules[display_banner]"
+                                               value="every">
+                                        <i class="input-helper"></i>
+                                    </label>
+                                </div>
+                            </div>
+
+                            <div class="col-md-2 col-sm-2 p-l-0">
+                                <input v-model="displayBannerEvery"
+                                       type="text"
+                                       name="pageview_rules[display_banner_every]"
+                                       class="form-control fg-input every-input"
+                                       id="num">
+                            </div>
+
+                            <div class="col-md-4 col-sm-4 p-l-0" style="height: 38px; line-height: 38px;">
+                                page views
+                            </div>
+                        </div>
                     </div>
                 </div>
+            </div>
+        </div>
 
-
-                <div class="col-md-2 col-sm-2 col-xs-2">
-                    <span class="btn btn-sm palette-Cyan bg waves-effect" v-on:click.prevent="addRule"><i class="zmdi zmdi-plus-circle"></i></span>
+        <div class="input-group fg-float m-t-20">
+            <div class="col-md-5 col-sm-5 p-r-0" style="margin-left: -15px;">
+                <div class="input-group fg-float m-t-10 checkbox">
+                    <label class="m-l-15">
+                        Display to user
+                        <input v-model="displayTimes"
+                               type="checkbox"
+                               name="pageview_rules[display_times]">
+                        <i class="input-helper"></i>
+                    </label>
                 </div>
             </div>
+
+            <div class="col-md-2 col-sm-2 p-l-0" style="margin-left: -10px;">
+                <input v-model="displayNTimes"
+                       type="text"
+                       name="pageview_rules[display_n_times]"
+                       class="form-control fg-input every-input"
+                       id="times_num">
+            </div>
+
+            <div class="col-md-4 col-sm-4 p-l-0" style="height: 38px; line-height: 38px;">
+                times, then stop.
+            </div>
+        </div>
+
+        <div class="input-group fg-float m-t-10 checkbox">
+            <label class="m-l-15">
+                Display once per session
+                <input v-model="oncePerSessionVal" value="1" name="once_per_session" type="checkbox">
+                <i class="input-helper"></i>
+            </label>
         </div>
     </div>
 </template>
@@ -51,48 +96,72 @@
         vertical-align: top;
         padding-top: 10px;
     }
+    .every-input {
+        height: 38px;
+        text-align: center;
+    }
 </style>
 
 <script type="text/javascript">
-    import vSelect from "remp/js/components/vSelect.vue";
-
     export default {
-        components: {
-            vSelect
-        },
         props: [
-            "pageviewRules"
+            "pageviewRules",
+            "oncePerSession"
         ],
         data() {
             return {
-                pageviewRulesData: null,
-                pageviewRulesOptions: [
-                    {"label": "Every", "value": "every"},
-                    {"label": "Since", "value": "since"},
-                    {"label": "Before", "value": "before"}
-                ]
+                displayBanner: 'always',
+                displayBannerEvery: 2,
+                displayTimes: false,
+                displayNTimes: 2,
+                oncePerSessionVal: false,
             };
         },
         created: function () {
-            this.pageviewRulesData = this.pageviewRules;
+            if (this.pageviewRules.display_banner !== undefined) {
+                this.displayBanner = this.pageviewRules.display_banner;
+            }
+            if (this.pageviewRules.display_banner_every !== undefined) {
+                this.displayBannerEvery = this.pageviewRules.display_banner_every;
+            }
+            if (this.pageviewRules.display_times !== undefined) {
+                this.displayTimes = this.pageviewRules.display_times;
+            }
+            if (this.pageviewRules.display_n_times !== undefined) {
+                this.displayNTimes = this.pageviewRules.display_n_times;
+            }
 
-            if (!this.pageviewRulesData.length) this.addRule();
+            this.oncePerSessionVal = this.oncePerSession;
+            this.updatePageviewRules();
+        },
+        watch: {
+            displayBanner: function () {
+                this.updatePageviewRules();
+            },
+            displayBannerEvery: function () {
+                this.updatePageviewRules();
+            },
+            displayTimes: function () {
+                this.updatePageviewRules();
+            },
+            displayNTimes: function () {
+                this.updatePageviewRules();
+            },
+            oncePerSessionVal: function () {
+                this.updatePageviewRules();
+            },
         },
         methods: {
-            addRule() {
-                this.pageviewRulesData.push({
-                    rule: null,
-                    num: null,
-                })
-            },
-            removeRule(i) {
-                var self = this;
-
-                this.pageviewRulesData.splice(i, 1);
-
-                if (!this.pageviewRulesData.length) {
-                    this.addRule();
-                }
+            updatePageviewRules: function () {
+                this.$emit('pageviewRulesModified', {
+                    rules: {
+                        display_banner: this.displayBanner,
+                        display_banner_every: this.displayBannerEvery,
+                        display_times: this.displayTimes,
+                        display_n_times: this.displayNTimes
+                    },
+                    oncePerSession: this.oncePerSessionVal
+                });
             }
         }
     }